pyproject.toml
[project]
name = "mnist-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.1",
    "pillow>=12.1.0",
    "plotly>=6.5.2",
    "scikit-learn>=1.8.0",
    "streamlit>=1.53.1",
]
